# docs/swagger.yaml
basePath: /
definitions:
  admin.AuthLogin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  admin.RegisterAdmin:
    properties:
      email:
        type: string
      fullname:
        type: string
      no_hp:
        type: string
      password:
        type: string
    required:
    - email
    - fullname
    - no_hp
    - password
    type: object
  admin.UpdateCustomer:
    properties:
      email:
        type: string
      fullname:
        type: string
      id:
        type: integer
      no_hp:
        type: string
      password:
        type: string
      pin:
        type: integer
    required:
    - email
    - fullname
    - id
    type: object
...
  response.Error:
    properties:
      code:
        type: integer
      messages:
        type: string
    type: object
  response.Result:
    properties:
      code:
        type: integer
      messages:
        type: string
      result: {}
    type: object
host: api-poins-id.herokuapp.com/v1
info:
  contact: {}
  description: Berikut API Poins ID
  title: API Poins ID
  version: "1.0"
paths:
  /admin:
    get:
      consumes:
      - application/json
      description: Dashboard for admin
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: Dashboard Admin
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: create admin with data
      parameters:
      - description: admin
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/admin.RegisterAdmin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: Create admin
      tags:
      - Admin
  /admin/approve/{transactionid}:
    post:
      consumes:
      - application/json
      description: Approve Transaction
      parameters:
      - description: transaction_id
        in: path
        name: transactionid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: Approve Transaction
      tags:
      - Admin
  /admin/customer:
    delete:
      consumes:
      - application/json
      description: Delete Customers for admin
      parameters:
      - description: id customer
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: Delete Customers
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: Find customers for admin
      parameters:
      - description: pagination
        in: query
        name: limit
        required: true
        type: integer
      - description: pagination
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: Find customers
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Transaction By Date for admin
      parameters:
      - description: admin
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateCustomer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: Transaction By Date
      tags:
      - Admin
  /admin/customer/point:
    put:
      consumes:
      - application/json
      description: Update customer point for admin
      parameters:
      - description: id customer
        in: query
        name: id
        required: true
        type: integer
      - description: point customer
        in: query
        name: point
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update customer point
      tags:
      - Admin
  /admin/history:
    get:
      consumes:
      - application/json
      description: Find History Customers for admin
      parameters:
      - description: pagination
        in: query
        name: limit
        required: true
        type: integer
      - description: pagination
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: Find History Customers
      tags:
      - Admin
  /admin/historystore:
    get:
      consumes:
      - application/json
      description: History Store for admin
      parameters:
      - description: pagination
        in: query
        name: limit
        required: true
        type: integer
      - description: pagination
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: History Store
      tags:
      - Admin
  /admin/login:
    post:
      consumes:
      - application/json
      description: Login admin
      parameters:
      - description: admin
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/admin.AuthLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: Login admin
      tags:
      - Admin
  /admin/stock:
    get:
      consumes:
      - application/json
      description: Stock Product for admin
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: Stock Product
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update Stock Product for admin
      parameters:
      - description: id product
        in: query
        name: id
        required: true
        type: integer
      - description: balance product
        in: query
        name: balance
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update Stock Product
      tags:
      - Admin
  /admin/store:
    delete:
      consumes:
      - application/json
      description: Delete Store for admin
      parameters:
      - description: id store
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: Delete Store
      tags:
      - Admin
  /admin/transaction:
    get:
      consumes:
      - application/json
      description: Transaction By Date for admin
      parameters:
      - description: "2022-01-01"
        in: query
        name: start
        type: string
      - description: "2022-12-31"
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: Transaction By Date
      tags:
      - Admin
  /admin/transaction/pending:
    get:
      consumes:
      - application/json
      description: Transaction Pending for Admin
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: Transaction Pending
      tags:
      - Admin
...
swagger: "2.0"